keep version info
