+++ internal/testplanet/planet.go
@@ -312,34 +312,3 @@ func (planet *Planet) WriteWhitelist(version storj.IDVersion) (string, error) {
 
 	return whitelistPath, err
 }
-
-// newVersionControlServer initializes the Versioning Server
-func (planet *Planet) newVersionControlServer() (peer *versioncontrol.Peer, err error) {
-
-	prefix := "versioncontrol"
-	log := planet.log.Named(prefix)
-	dbDir := filepath.Join(planet.directory, prefix)
-
-	if err := os.MkdirAll(dbDir, 0700); err != nil {
-		return nil, err
-	}
-
-	config := &versioncontrol.Config{
-		Address: "127.0.0.1:0",
-		Versions: versioncontrol.ServiceVersions{
-			Satellite:   "v0.0.1",
-			Storagenode: "v0.0.1",
-			Uplink:      "v0.0.1",
-			Gateway:     "v0.0.1",
-			Identity:    "v0.0.1",
-		},
-	}
-	peer, err = versioncontrol.New(log, config)
-	if err != nil {
-		return nil, err
-	}
-
-	log.Debug(" addr= " + peer.Addr())
-
-	return peer, nil
-}
